Refactor url to point to a more generic weather station
The commit refactors the specified URL in the modules/gui/xmobar.nix
file to direct users to a more general weather station website. The
change entails adjusting the openUrlCmd function to link to
"https://www.meteo.pl/" instead of the previous specific URL. This
modification aims to enhance the flexibility and usability of the script
by providing a broader resource for weather information. By updating the
URL to a more generic location, users have access to a wider range of
weather data, improving the overall functionality and relevance of the
script in varying contexts.