Add script to connect to BinarApps VPN
This commit adds a script in the `modules/security/default.nix` file to
simplify connecting to the BinarApps VPN. The script utilizes `sudo
openvpn` command with the file `binarapps-office.ovpn`. This change
enhances the user experience by providing an easy and straightforward
way to establish a VPN connection to BinarApps.