Use latest configuration for binarapps vpn
This commit implements the use of the most recent configuration settings
for the Binarapps VPN. The changes involve updating the
`modules/security/binarapps-office.ovpn` file with the new configuration
options. The modifications include specifying the protocol as UDP,
setting the remote IP address to `51.75.40.167`, defining the port as
`1194`, adjusting encryption settings, and filtering certain pull
options.