Add pandoc document converter to module utils/default.nix
This commit adds the pandoc document converter to the list of packages
in the utils module. Pandoc is a versatile tool for converting between
various document formats, which aligns with the objective of enhancing
the utility of the Nix packages provided in this module. By including
pandoc, users will have access to a powerful document conversion tool,
expanding the capabilities of the module.