Add a missing yt-dlp downloader
This commit includes the addition of the yt-dlp downloader under the
'programs' section of the 'modules/utils/default.nix' file. The 'yt-dlp'
package is enabled to ensure its availability for use within the system
configuration. This addition enhances the utility packages offered by
the system, providing users with the capability to utilize yt-dlp for
downloading content efficiently.