Add freetube to provide an alternative for watching YouTube videos
The commit includes the insertion of the freetube utility into the
default.nix file within the list of utilities. This addition aims to
alleviate the irritations users face while watching content on the
YouTube website, offering them an improved viewing experience through
the utilization of freetube as a YouTube client. By integrating freetube
into the module, users are granted a more pleasant and tailored
experience for watching YouTube videos.