Update NixOS and Home Manager to 24.05
This commit updates the version of Home Manager to 24.05 within the
Makefile and the system state version for the Omega machine to 24.05 in
the configuration.nix file. Additionally, changes have been made in the
default.nix file for the security module. The pinentryPackage is set to
pkgs.pinentry-gtk2 to ensure compatibility. Lastly, in the utils module,
the minipro EEPROM programmer has been commented out to prevent its
inclusion. These modifications collectively align the system and utility
configurations with the updated versions for improved functionality and
security.