Enable splitting vimrc into a distinct module
This commit adjusts the editor default.nix file to enable the separation
of the vimrc configuration into its standalone module. By introducing
the editor.RCs option, a list of vimrc configuration strings can now be
included. This modification allows for a more organized and modular
setup of the vimrc configurations. With this change, the vimrc settings
are now managed separately, enhancing flexibility and customization
options for the vimrc configuration.