Make fugitive an autonomic module
This commit adjusts the editor default.nix file to enable the separation
of the fugitive configuration into its standalone module. By creating
the fugitive.nix file, the organization and modularity of the
configuration setup are improved. The fugitive.nix now contains the
necessary configurations that were previously part of the default.nix.
Through this segregation, the management and independence of the
fugitive configuration are enhanced.