Move GitHub CLI configuration to a separate module
This commit moves the GitHub CLI configuration from the default.nix file
to a new module named gh.nix. By organizing the configuration into
separate modules, the structure and maintenance of the VCS settings are
improved. The new gh.nix module contains the settings for GitHub CLI,
specifying parameters such as git protocol, editor, prompt, and aliases.
Additionally, the programs.gh-dash.enable setting is included within
this module. This reorganization enhances the clarity and modularity of
the VCS configuration setup.