Move utilities from shell module to utils
This commit restructures the project by transferring the utility
configurations from the `shell` module to the `utils` module. The
`nnn.nix` file originally located in the `shell` module has been renamed
and moved to the `utils` module. Additionally, changes were made to the
`default.nix` files in both the `shell` and `utils` modules. The
`nnn.nix` file is now imported and included in the configurations within
the `utils` module while being removed from the `shell` module.
Furthermore, configurations for `fzf`, `lsd`, and `zoxide` utilities
were relocated from the `shell` module to the `utils` module to
consolidate all utility settings. This restructuring enhances the
organization and maintainability of the VCS setup, making it easier to
manage and update utility configurations effectively.