Remove unnecessary file
The file "speak.nix" from the path "modules/utils/speak.nix" has been
removed as it was deemed unnecessary for the project. This file
contained scripts for handling text-to-speech functionalities, image
processing, video editing, audio management, and a script for generating
daily psalm numbers based on the date input. The content of this file
was no longer in use in the current iteration of the project and was
thus deleted to declutter the repository.

The removal of this file ensures that the project remains focused on the
essential functionalities required for its current scope without
unnecessary scripts that are not actively utilized.