Update editor and utility plugins configuration
The changes in this commit involve modifications to the editor plugins
and utility configuration files. In the "fugitive.nix" file, the
function for opening a pull request has been updated to call the
'tertiusPullRequestWindow' function. Additionally, in the "tertius.nix"
file, the function name related to pull request description has been
changed to 'tertiusPullRequestWindow'. Furthermore, in the "autocmd.vim"
file, the mapping for cherry-picking to a branch has been updated to use
the function 'GitCherryPickToBranchFromLine'. Moreover, in the
"mappings.vim" file, new mapping commands have been added for
Git-related operations. Lastly, the "tertius.nix" utility file has been
modified to refine the 'user_story_id_from_branch' function. These
changes improve the workflow and enhance the usability of the editor and
utility plugins, aligning them with the project requirements.