Add functionality to open a file in git hosting context of the branch
The modifications in this commit include the addition of a new feature
to open a file from the editor in a Git hosting platform within the
context of the current branch. The changes are made to the
"fugitive.nix" plugin file within the editor plugins module. Several new
functions have been introduced for extracting project details like the
project name, owner, and the current branch. Additionally, the
'gitHostingUrl' function has been implemented to generate the hosting
URL specific to the current project, owner, branch, and file. This
commit also introduces a mapping 'nnoremap' command for triggering the
opening of the file on the Git hosting platform using the configured
browser. These enhancements provide a convenient way to directly view
files in the context of the branch within the selected Git hosting
service.