Add Discord to GUI module dependencies
This commit includes the addition of Discord to the list of GUI module
dependencies in the `default.nix` file. This change responds to a user
request to integrate Discord into the environment setup, allowing users
to easily install and access it alongside other applications. The
inclusion of Discord facilitates better communication and collaboration,
enhancing the overall user experience within the GUI framework.