Enhance utilization of projectsDirectory setting
This commit modifies the use of the `projectsDirectory` setting in both
the shell aliases and utility configurations. Specifically, the
hardcoded path to the Projects directory is replaced with
`${config.projectsDirectory}` in `modules/shell/aliases.nix`, ensuring
that the aliases are now dynamic and can adapt to any changes in the
configuration file. Additionally, in `modules/utils/nnn.nix`, the path
for bookmarks is updated to reference `config.projectsDirectory` instead
of a static path of `~/Projects`. These changes create a more flexible
configuration that can respond to user customization, adhering to the
requirements in the user story for improved configurability of directory
paths.