Add a service to synchronize projects
This commit introduces a new service and timer to synchronize the user's
projects with a cloud server using `rsync`. A new file,
`modules/utils/sync_projects.nix`, has been created to define a systemd
user service that executes the `rsync` command, ensuring that the
projects in the specified directory are backed up to the cloud at a
defined interval. Furthermore, the `default.nix` file has been updated
to import this new functionality and include an option for the
`cloudDomain`. These changes enhance the project's configurability and
automate the backup process, addressing the user story's requirement for
a synchronization service.