Reorder mappings in mappings.vim
This commit reorganizes the key mappings defined in
`modules/editor/post/mappings.vim` to improve their logical grouping and
usability. The prior ordering of the mappings did not follow a
consistent structure, which could lead to confusion for users trying to
remember or locate specific commands. The mappings for the pull request
and issue window have been switched to create a more intuitive flow.
Overall, this change enhances the readability and accessibility of the
mappings, facilitating a better user experience while using the editor
features.