Add tools for manipulating todo files
This commit introduces functions and mappings to enhance the management
of todo items within the `todo.nix` and `autocmd.vim` files. The main
feature is the addition of a `todoToggleItem` function, which allows
users to easily mark todo items as completed or not by toggling their
states between `[ ]` and `[x]`. Additionally, a new key mapping is
established for this function within the xit file type, enabling swift
interaction with todo items. These changes address the need for improved
todo item manipulation as outlined in the user story, streamlining the
workflow for users managing their tasks.